changed all double to int

+++ src/main/java/agh/ics/oop/Comparators.java
@@ -7,7 +7,7 @@ import java.util.Comparator;
 class EnergyComparator implements Comparator<Animal> {
     @Override
     public int compare(Animal animal0, Animal animal1){
-        return Double.compare(animal1.getEnergy(), animal0.getEnergy());
+        return Integer.compare(animal1.getEnergy(), animal0.getEnergy());
     }
 }
 class ComparatorX implements Comparator<IMapElement> {
